requirements.txt
tensorflow-intel==2.10.0
tensorboard==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
numpy==1.24.3
flask 
pillow 
requests 
matplotlib
opencv-python
